feat(matematicas): implementar resolución de sistemas de ecuaciones lineales NxN Closes #674
